+++ src/vuex/actions/forms.js
@@ -1,56 +1,8 @@
-import auth from '../api/auth'
-import recommendCompany from '../api/recommend-company'
-import recommendWorker from '../api/recommend-worker'
-import registerUser from '../api/register-user'
-import recommendations from '../api/recommendations'
-import {PERMISSIONS} from '../utils/translations.js'
+import recommendCompany from '../../api/recommend-company'
+import recommendWorker from '../../api/recommend-worker'
+import registerUser from '../../api/register-user'
 
-import * as types from './mutation-types'
-
-export const authorize = ({ dispatch }, user) => {
-  dispatch(types.AUTH_REQUEST)
-  auth.authorize(user).then(
-    (response) => {
-      dispatch(types.AUTH_SUCCESS, user, response.headers('x-auth-token'), response.data.permissions, response.data.name, response.data.surname)
-      dispatch(types.CHANGE_ROUTE, {path: '/app/home'})
-      dispatch(types.ALERT_SHOW, {
-        header: 'Sukces!',
-        type: 'success',
-        body: 'Użytkownik został zalogowany'
-      })
-
-      auth.availablePermissions().then(({ data }) => {
-        const allPerms = Object.keys(data).map(key => ({
-          name: PERMISSIONS[data[key]],
-          value: data[key]
-        }))
-
-        dispatch(types.RECEIVE_ALL_PERMISSIONS, allPerms)
-      })
-    },
-    () => {
-      dispatch(types.AUTH_FAILURE)
-      dispatch(types.ALERT_SHOW, {
-        header: 'Błąd!',
-        type: 'danger',
-        body: 'Błąd podczas logowania'
-      })
-    }
-  )
-  setTimeout(() => dispatch(types.ALERT_HIDE), 3000)
-}
-
-export const invalidate = ({ dispatch }) => {
-  dispatch(types.AUTH_REQUEST)
-  dispatch(types.AUTH_INVALIDATE)
-  dispatch(types.CHANGE_ROUTE, { path: '/login' })
-  dispatch(types.ALERT_SHOW, {
-    header: 'Sukces!',
-    type: 'success',
-    body: 'Użytkownik został wylogowany'
-  })
-  setTimeout(() => dispatch(types.ALERT_HIDE), 3000)
-}
+import * as types from '../mutation-types'
 
 export const sendRegisterUserForm = ({ dispatch }, form) => {
   dispatch(types.FORM_REGISTER_USER_REQUEST, form)
@@ -167,10 +119,4 @@ export const sendWorkerRecommendationForm = ({ dispatch }, form, fileData) => {
       })
     }
   )
-}
-
-export const getAllRecommendations = ({ dispatch }) => {
-  recommendations.getAll().then(res => {
-    dispatch(types.RECEIVE_RECOMMENDATIONS, res.data)
-  })
 }
