+++ src/vuex/actions/recommendations.js
@@ -1,0 +1,9 @@
+import recommendations from '../../api/recommendations'
+
+import * as types from '../mutation-types'
+
+export const getAllRecommendations = ({ dispatch }) => {
+  recommendations.getAll().then(res => {
+    dispatch(types.RECEIVE_RECOMMENDATIONS, res.data)
+  })
+}
